+++ init/main.c
@@ -857,7 +857,7 @@ static void __init do_initcalls(void)
 {
 	int level;
 
-	for (level = 0; level < ARRAY_SIZE(initcall_levels) - 1; level++)
+	for (level = 0; level < (int)ARRAY_SIZE(initcall_levels) - 1; level++)
 		do_initcall_level(level);
 }
